+++ .dojorc
@@ -9,6 +9,7 @@
 			"src/accordion-pane",
 			"src/avatar",
 			"src/breadcrumb",
+			"src/breadcrumb-group",
 			"src/button",
 			"src/calendar",
 			"src/card",
