+++ src/breadcrumb/tests/Breadcrumb.spec.tsx
@@ -1,92 +1,46 @@
 const { describe, it } = intern.getInterface('bdd');
-import assertionTemplate from '@dojo/framework/testing/assertionTemplate';
 import harness from '@dojo/framework/testing/harness';
 import { tsx } from '@dojo/framework/core/vdom';
 
 import * as css from '../../theme/default/breadcrumb.m.css';
-import Crumb from '../Crumb';
-import * as fixedCss from '../styles/breadcrumb.m.css';
 import Breadcrumb from '../index';
 
 describe('Breadcrumb', () => {
-	const textSeparatorItemStyles = [fixedCss.listItemFixed, css.listItem, undefined];
-	const currentTextSeparatorItemStyles = [fixedCss.listItemFixed, css.listItem, css.current];
+	it('renders a text label', () => {
+		const h = harness(() => <Breadcrumb item={{ label: 'Home' }} />);
 
-	const baseTemplate = assertionTemplate(() => (
-		<nav assertion-key="root" classes={css.root} aria-label="breadcrumb">
-			<ol classes={[fixedCss.listFixed, css.list]}>
-				<li classes={textSeparatorItemStyles} key="breadcrumb-0">
-					<Crumb item={{ label: 'Home' }} current={undefined} />
-				</li>
-				<li classes={textSeparatorItemStyles} key="breadcrumb-1">
-					<Crumb item={{ label: 'Overview' }} current={undefined} />
-				</li>
-				<li classes={currentTextSeparatorItemStyles} key="breadcrumb-2">
-					<Crumb item={{ label: 'Tests' }} current="page" />
-				</li>
-			</ol>
-		</nav>
-	));
+		h.expect(() => (
+			<span aria-current={undefined} classes={css.root}>
+				Home
+			</span>
+		));
+	});
 
-	it('renders with items', () => {
+	it('renders with a link', () => {
 		const h = harness(() => (
 			<Breadcrumb
-				label="breadcrumb"
-				items={[{ label: 'Home' }, { label: 'Overview' }, { label: 'Tests' }]}
+				item={{
+					href: '/home',
+					label: 'Home',
+					title: 'Home page'
+				}}
 			/>
 		));
 
-		h.expect(baseTemplate);
+		h.expect(() => (
+			<a aria-current={undefined} classes={css.root} href="/home" title="Home page">
+				Home
+			</a>
+		));
 	});
 
-	it('renders with a custom child renderer', () => {
-		const h = harness(() => (
-			<Breadcrumb
-				label="breadcrumb"
-				items={[
-					{ label: 'Home', href: '/' },
-					{ label: 'Overview', href: '/overview', title: 'Breadcrumb Overview' },
-					{ href: '/tests', label: 'Tests' }
-				]}
-			>
-				{(items) => (
-					<ol>
-						{items.map((item, i) => (
-							<li key={`crumb-${i}`}>
-								<a
-									aria-current={i === 1 ? 'page' : undefined}
-									href={item.href}
-									title={item.title}
-								>
-									{item.label}
-								</a>
-							</li>
-						))}
-					</ol>
-				)}
-			</Breadcrumb>
-		));
+	it('renders with an aria-current attributes', () => {
+		const h = harness(() => <Breadcrumb current="step" item={{ label: 'Home' }} />);
 
-		h.expect(
-			baseTemplate.setChildren('~root', () => [
-				<ol>
-					<li key="crumb-0">
-						<a aria-current={undefined} href="/" title={undefined}>
-							Home
-						</a>
-					</li>
-					<li key="crumb-1">
-						<a aria-current="page" href="/overview" title="Breadcrumb Overview">
-							Overview
-						</a>
-					</li>
-					<li key="crumb-2">
-						<a aria-current={undefined} href="/tests" title={undefined}>
-							Tests
-						</a>
-					</li>
-				</ol>
-			])
-		);
+		h.expect(() => (
+			<span aria-current="step" classes={css.root}>
+				Home
+			</span>
+		));
 	});
 });
