+++ src/theme/default/breadcrumb-group.m.css
@@ -1,0 +1,16 @@
+.root {
+}
+
+.list {
+}
+
+.listItem,
+.listItem::before {
+	margin-right: 0.75rem;
+}
+.listItem::last-child {
+	margin-right: 0;
+}
+
+.current {
+}
