+++ src/theme/default/breadcrumb.m.css
@@ -1,16 +1,2 @@
 .root {
 }
-
-.list {
-}
-
-.listItem,
-.listItem::before {
-	margin-right: 0.75rem;
-}
-.listItem::last-child {
-	margin-right: 0;
-}
-
-.current {
-}
