+++ src/theme/default/breadcrumb.m.css.d.ts
@@ -1,4 +1,1 @@
 export const root: string;
-export const list: string;
-export const listItem: string;
-export const current: string;
