--- src/theme/default/crumb.m.css
@@ -1,2 +1,0 @@
-.root {
-}
