--- src/theme/default/crumb.m.css.d.ts
@@ -1,1 +1,0 @@
-export const root: string;
