+++ src/theme/default/index.ts
@@ -1,6 +1,7 @@
 import * as accordionPane from './accordion-pane.m.css';
 import * as avatar from './avatar.m.css';
 import * as breadcrumb from './breadcrumb.m.css';
+import * as breadcrumbGroup from './breadcrumb-group.m.css';
 import * as button from './button.m.css';
 import * as calendar from './calendar.m.css';
 import * as card from './card.m.css';
@@ -8,7 +9,6 @@ import * as checkbox from './checkbox.m.css';
 import * as checkboxGroup from './checkbox-group.m.css';
 import * as chip from './chip.m.css';
 import * as constrainedInput from './constrained-input.m.css';
-import * as crumb from './crumb.m.css';
 import * as dialog from './dialog.m.css';
 import * as emailInput from './email-input.m.css';
 import * as grid from './grid.m.css';
@@ -51,7 +51,7 @@ export default {
 	'@dojo/widgets/accordion-pane': accordionPane,
 	'@dojo/widgets/avatar': avatar,
 	'@dojo/widgets/breadcrumb': breadcrumb,
-	'@dojo/widgets/breadcrumb/Crumb': crumb,
+	'@dojo/widgets/breadcrumb-group': breadcrumbGroup,
 	'@dojo/widgets/button': button,
 	'@dojo/widgets/calendar': calendar,
 	'@dojo/widgets/card': card,
