+++ src/theme/material/breadcrumb-group.m.css.d.ts
@@ -1,0 +1,4 @@
+export const root: string;
+export const list: string;
+export const listItem: string;
+export const current: string;
