+++ src/theme/material/index.ts
@@ -1,6 +1,6 @@
 import * as accordionPane from './accordion-pane.m.css';
 import * as avatar from './avatar.m.css';
-import * as breadcrumb from './breadcrumb.m.css';
+import * as breadcrumbGroup from './breadcrumb-group.m.css';
 import * as button from './button.m.css';
 import * as card from './card.m.css';
 import * as calendar from './calendar.m.css';
@@ -49,7 +49,7 @@ import * as tooltip from './tooltip.m.css';
 export default {
 	'@dojo/widgets/accordion-pane': accordionPane,
 	'@dojo/widgets/avatar': avatar,
-	'@dojo/widgets/breadcrumb': breadcrumb,
+	'@dojo/widgets/breadcrumb-group': breadcrumbGroup,
 	'@dojo/widgets/button': button,
 	'@dojo/widgets/calendar': calendar,
 	'@dojo/widgets/checkbox-group': checkboxGroup,
